+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 6.1-SNAPSHOT
 ADDED:
 
 - new method that can delete files recursively with a given prefix in a source directory 
+- new unit test classes for the class SystemFileExtensions and SystemPropertiesExtensions created
 
 CHANGED: